scripts/requirements.txt
web3>=6.0.0
psycopg2-binary>=2.9.0
